requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Data download
ccxt>=4.0.0
requests>=2.31.0

# Technical analysis
ta>=0.11.0

# Machine learning / preprocessing
scikit-learn>=1.3.0

# Deep learning
torch>=2.1.0
pytorch-lightning>=2.1.0
pytorch-forecasting>=1.0.0

# Hyperparameter optimization
optuna>=3.4.0

# Experiment tracking
tensorboard>=2.15.0

# Progress bars
tqdm>=4.65.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# API serving
fastapi>=0.110.0
uvicorn>=0.27.0
